+++ demos/world/src/test/java/is/codion/framework/demos/world/model/CountryTableModelTest.java
@@ -23,7 +23,7 @@ import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.demos.world.domain.WorldImpl;
-import is.codion.framework.demos.world.domain.api.World;
+import is.codion.framework.demos.world.domain.api.World.Country;
 import is.codion.swing.common.model.worker.ProgressWorker.ProgressReporter;
 
 import org.junit.jupiter.api.Test;
@@ -37,7 +37,7 @@ public final class CountryTableModelTest {
 	void fillCountryReport() throws ReportException {
 		try (EntityConnectionProvider connectionProvider = createConnectionProvider()) {
 			CountryTableModel tableModel = new CountryTableModel(connectionProvider);
-			tableModel.conditionModel().conditionModel(World.Country.CODE).setEqualValue("ISL");
+			tableModel.conditionModel().conditionModel(Country.CODE).equalValue().set("ISL");
 			tableModel.refresh();
 			tableModel.selectionModel().setSelectedIndex(0);
 			tableModel.fillCountryReport(new ProgressReporter<String>() {