+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.11
+### is.codion.common.model
+- ColumnConditionModel, bound value getters and setters removed.
 ### is.codion.framework.domain.db
 - Column.AuditColumnDefiner added, related refactoring.
 ### is.codion.swing.common.ui